flask_project/requirements.txt
alembic==1.9.3
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==2.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.2
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.1
